+++ Binary Tree Exercise/binary_tree.c
@@ -3,19 +3,19 @@
 #include "tree.h"
 
 // Creates the tree:
-void create(TREE *tree) {
+void create_tree(TREE *tree) {
   *tree = NULL;
 }
 
 // Searches the tree for an element:
-int search(ITEM *x, POINT *P) {
+int search_element(ITEM *x, POINT *p) {
   if (*p == NULL) return -1;
 
   if (x->key < (*p)->item.key)
-    return search(x, &(*p)->left);
+    return search_element(x, &(*p)->left);
 
   if (x->key > (*p)->item.key)
-    return search(x, &(*p)->left);
+    return search_element(x, &(*p)->left);
 
   *x = (*p)->item;
 
@@ -23,9 +23,9 @@ int search(ITEM *x, POINT *P) {
 }
 
 // Inserts an element into the tree:
-int insert(ITEM x, POINT *p) {
+int insert_element(ITEM x, POINT *p) {
   if (*p == NULL) {
-    *p = (PONT) malloc(sizeof(NODE));
+    *p = (POINT) malloc(sizeof(NODE));
 
     (*p)->item = x;
     (*p)->left = NULL;
@@ -35,11 +35,11 @@ int insert(ITEM x, POINT *p) {
   }
 
   if (x.key < (*p)->item.key) {
-    return insert(x, &(*p)->left);
+    return insert_element(x, &(*p)->left);
   }
 
   if (x.key > (*p)->item.key) {
-    return insert(x, &(*p)->right);
+    return insert_element(x, &(*p)->right);
   }
 
   return -1;
@@ -60,23 +60,23 @@ void predecessor(POINT q, POINT *r) {
 }
 
 // Removes an element from the tree:
-int remove(ITEM x, POINT *p) {
+int remove_element(ITEM x, POINT *p) {
   POINT aux;
 
   if (*p == NULL)
     return -1;
 
-    if (x->key < (*p)->item.key)
-      return remove(x, &(*p)->left);
+    if (x.key < (*p)->item.key)
+      return remove_element(x, &(*p)->left);
 
-    if (x->key > (*p)->item.key)
-      return remove(x, &(*p)->left);
+    if (x.key > (*p)->item.key)
+      return remove_element(x, &(*p)->right);
 
     if ((*p)->right == NULL) {
       aux = *p;
       *p = (*p)->left;
       free(aux);
-      return 0
+      return 0;
     }
 
     if ((*p)->left == NULL) {
@@ -86,6 +86,30 @@ int remove(ITEM x, POINT *p) {
       return 0;
     }
 
-    predecessor(*p, &(*p)-left);
+    predecessor(*p, &(*p)->left);
     return 0;
+}
+
+void go_through_in(POINT p) {
+  if (p != NULL) {
+    go_through_in(p->left);
+    printf("%d\n", p->item.key);
+    go_through_in(p->right);
+  }
+}
+
+void go_through_pre(POINT p) {
+  if (p != NULL) {
+    printf("%d\n", p->item.key);
+    go_through_pre(p->left);
+    go_through_pre(p->right);
+  }
+}
+
+void go_through_pos(POINT p) {
+  if (p != NULL) {
+    go_through_pos(p->left);
+    go_through_pos(p->right);
+    printf("%d\n", p->item.key);
+  }
 }